feat: main-page header
