+++ src/components/header/Header.tsx
@@ -1,0 +1,28 @@
+import "./Header.scss";
+
+export default function Header() {
+  return (
+    <>
+      <header>
+        <nav>
+          <div className="container">
+            <div className="d-flex">
+              <img src="/img/main/logo.png" alt="logo" />
+              <div>
+                <a href="#firstPage">
+                  <span>공지사항</span>
+                </a>
+                <a href="#secondPage">
+                  <span>질문게시판</span>
+                </a>
+                <a href="#thirdPage">
+                  <span>건의게시판</span>
+                </a>
+              </div>
+            </div>
+          </div>
+        </nav>
+      </header>
+    </>
+  );
+}
